% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pal.R
\name{col_bg}
\alias{col_bg}
\alias{col_fg}
\title{Title}
\usage{
col_bg(theme = NA, fade = 0)

col_fg(theme = NA, fade = 0)
}
\arguments{
\item{theme}{A string corresponding to one of the 31 themes available in RStudio (see \code{\link{list_pal}}).
If not specified, the function tries to retrieve the active theme using \pkg{rstudioapi}.}

\item{fade}{A numeric value from 0 to 1. How much the color should be faded (in the background/foreground color).}
}
\value{
A color (hexadecimal format).
}
\description{
Title
}
