% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pal.R
\name{get_pal}
\alias{get_pal}
\alias{list_pal}
\alias{viz_pal}
\title{Color palettes matching RStudio themes}
\usage{
get_pal(theme = NA, n)

list_pal()

viz_pal(pal, print.ribbon = TRUE, print.hex = FALSE)
}
\arguments{
\item{theme}{Theme of the palette. A string corresponding to one of the 31 themes available in RStudio (see Details).
If not specified, the function tries to retrieve the active theme using \pkg{rstudioapi}.}

\item{n}{Number of colors to return. Currently limited to 5.}

\item{pal}{A color palette as returned by \code{get_pal}.}

\item{print.ribbon}{Logical stating whether a ribbon with the name of the palette should be added to the plot (default \code{TRUE}).}

\item{print.hex}{Logical stating whether the hexadecimal code of each color should be added to the plot (default \code{FALSE}).}
}
\value{
A vector of hexadecimal colors with additional attributes.
}
\description{
The function \code{get_pal} generates a palette of colors according to a provided theme.
The list of available themes is returned by \code{list_pal} (see also Details).
Palettes can be previewed using the function \code{viz_pal}.
}
\details{
The available themes are:
\code{Ambiance}, \code{Chaos}, \code{Chrome}, \code{Clouds}, \code{Clouds Midnight}, \code{Cobalt},
\code{Crimson Editor}, \code{Dawn}, \code{Dracula}, \code{Dreamweaver}, \code{Eclipse}, \code{Idle Fingers},
\code{Katzenmilch}, \code{Kr Theme}, \code{Material}, \code{Merbivore}, \code{Merbivore Soft},
\code{Mono Industrial}, \code{Monokai}, \code{Pastel On Dark}, \code{Solarized Dark}, \code{Solarized Light},
\code{TextMate}, \code{Tomorrow}, \code{Tomorrow Night}, \code{Tomorrow Night 80s}, \code{Tomorrow Night Blue},
\code{Tomorrow Night Bright}, \code{Twilight}, \code{Vibrant Ink} and \code{Xcode}.
}
\note{
The function \code{viz_pal} is directly inspired by the function \code{print.palette}
available in the package \pkg{wesanderson} by Karthik Ram.
}
\examples{
\dontrun{
get_pal()
}

list_pal()
my_pal <- get_pal(theme = "Twilight")
viz_pal(my_pal, print.hex = TRUE)

par(mfrow = c(7, 5))
for(i in list_pal()) {viz_pal(get_pal(i))}

}
