% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pal.R
\name{print.edipal}
\alias{print.edipal}
\title{Print method for color palette}
\usage{
\method{print}{edipal}(x, ...)
}
\arguments{
\item{x}{the palette to be printed.}

\item{...}{further arguments to be passed to or from other methods. They are ignored in this function.}
}
\description{
Print method for color palette
}
