% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_base_sty.R
\name{set_base_sty}
\alias{set_base_sty}
\title{Set graphical parameters}
\usage{
set_base_sty(theme = NA, variation = 2)
}
\arguments{
\item{theme}{A string corresponding to one of the 31 themes available in RStudio (see \code{\link{list_pal}}).
If not specified, the function tries to retrieve the active theme using \pkg{rstudioapi}.}

\item{variation}{An integer ranging from 0 to 5 to select among 6 flavors for each theme.}
}
\description{
Set graphical parameters
}
\examples{

\dontrun{
set_base_sty()
}

set_base_sty("Tomorrow Night", 3)
hist(rnorm(100))
plot(iris$Sepal.Length, iris$Petal.Length)
barplot(1:5, names.arg = LETTERS[1:5])
boxplot(iris$Sepal.Length ~ iris$Species)
image(volcano)

par(mfrow = c(2, 3))
x <- rnorm(100)
for(i in list_pal()){
   for(j in 0:5){
       set_base_sty(i, variation = j)
       hist(x)
       title(paste(i, j))
   }
}

}
