% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pal.R
\name{col_bg}
\alias{col_bg}
\alias{col_fg}
\title{Get foreground and background colors}
\usage{
col_bg(x = NA, fade = 0)

col_fg(x = NA, fade = 0)
}
\arguments{
\item{x}{A color palette as returned by \code{get_pal} or a string corresponding to one of the 31 themes available in RStudio (see \code{\link{list_pal}}).
If NA, the function tries to retrieve the active theme using \pkg{rstudioapi}.}

\item{fade}{A numeric value from 0 to 1. How much the color should be faded (in the background/foreground color).}
}
\value{
A color (hexadecimal format).
}
\description{
Get foreground and background colors
}
