% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_theme.R
\name{theme_editor}
\alias{theme_editor}
\alias{scale_fill_editor}
\alias{scale_colour_editor}
\alias{scale_color_editor}
\title{Theme and color scales for \code{ggplot2}}
\usage{
theme_editor(theme = NA, base_size = 10, base_family = "sans")

scale_fill_editor(theme = NA, ...)

scale_colour_editor(theme = NA, ...)

scale_color_editor(theme = NA, ...)
}
\arguments{
\item{theme}{A string corresponding to one of the 31 themes available in RStudio (see \code{\link{list_pal}}).
If not specified, the function tries to retrieve the active theme using \pkg{rstudioapi}.}

\item{base_size}{base font size}

\item{base_family}{base font family}

\item{...}{Other arguments passed on to \code{\link{discrete_scale}}
to control name, limits, breaks, labels and so forth.}
}
\description{
Theme and color scales for \code{ggplot2}
}
\examples{

df <- data.frame(
gp = factor(rep(letters[1:3], each = 10)),
y = rnorm(30)
)

ggplot(df, aes(gp, y, color = gp, fill = gp)) +
  geom_boxplot(color = col_fg(fade = 0.2)) +
  labs(title = "A nice plot", subtitle = "This is a demo of edistyle") +
  facet_wrap(~gp) +
  theme_editor() +
  scale_fill_editor()

ggplot(iris, aes(Sepal.Length, Petal.Length)) +
  geom_point(color = col_fg(fade = 0.2)) +
  geom_smooth(color = get_pal()[1], fill = get_pal()[2]) +
  labs(title = "Edgar Anderson's Iris Data", subtitle = "This is a demo", caption = "edistyle v.0.1.0") +
  theme_editor() +
  scale_fill_editor()

}
